+++ frontend/src/components/ranking/Mjesto.css
