+++ frontend/src/components/scoreboard/Scoreboard.css
