+++ frontend/src/components/scoreboard/Scoreboard.jsx
