+++ frontend/src/pages/registration/Registracija.css
