+++ frontend/src/pages/registration/Registracija.jsx
