+++ frontend/src/pages/scoreboard/SBsemestar.jsx
