+++ frontend/src/pages/scoreboard/SBukupno.css
