+++ frontend/src/pages/scoreboard/SBukupno.jsx
