+++ frontend/src/pages/scoreboard/Scoreboard.css
