+++ frontend/src/pages/scoreboard/Scoreboard.jsx
